Backend/requirements.txt
flask==3.0.3
openai==1.42.0
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.111.1
uvicorn==0.25.0
replicate==0.12.0
python ==3.11